backend_python/pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "grpcio>=1.70.0",
    "protobuf>=7.34.0",
    "clickhouse-connect>=0.11.0",
    "clickhouse-sqlalchemy>=0.3.2",
    "insightface>=0.7.3",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.2",
    "opencv-python>=4.13.0.92",
    "psycopg2>=2.9.11",
    "textual>=8.0.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.19.0",
]
